requirements.txt
absl-py==2.3.1
astunparse==1.6.3
async-timeout==5.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
deepface==0.0.95
dotenv==0.9.9
filelock==3.18.0
fire==0.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
gdown==5.2.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.1.0
rsa==4.9.1
safetensors==0.6.1
six==1.17.0
soupsieve==2.8
spotipy==2.25.1
SQLAlchemy==2.0.42
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
